moved force kill tor to scripts
